requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
pypdf>=3.17.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
yt-dlp>=2024.0.0
openai-whisper>=20231117
